bpf: move ICMPv4 packet generation functions to a separate file
The goal is to make these functions reusable for the ICMP policy denial feature,
so moving them to a shared icmp.h file
